feat(macla): assign_retrospective_credit — TD-lambda over trace
Pure function: walks trace, computes TD-lambda(0.95)-weighted delta
per procedure, mutates (alpha, beta) on the memory system, returns
deltas dict for logging. Evicted proc_keys are silently skipped at
the mutation site but still appear in the returned dict so callers
can log the intended-but-unapplied credit.

+++ agents/macla/macla_lib.py
@@ -126,8 +126,8 @@ class Procedure:
     postconditions: list[str] = field(default_factory=list)
     reasoning: str = ""
     concepts: set[str] = field(default_factory=set)
-    alpha: int = 1
-    beta: int = 1
+    alpha: float = 1.0
+    beta: float = 1.0
     execution_count: int = 0
     generalisability_score: float = 0.5
     confidence: float = 0.5
@@ -152,8 +152,8 @@ class MetaProcedure:
     preconditions_meta: list[str]
     sub_procedures: list[str]
     composition_policy: dict[str, any]
-    alpha: int = 1
-    beta: int = 1
+    alpha: float = 1.0
+    beta: float = 1.0
     execution_count: int = 0
 
     @property
